+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added QueryParams class.
 - Added setQueryParams method to AbstractRequestMapperQuery class.
 - Added AbstractAliasTree class and AliasTreeInterface interface.
+- Added 'like' (lk / ~) and 'not like' (nl / !~) operators to FiltersQueryParam.
 
 ## 1.3.0 - 2023-09-17